+++ optional_plugins/loader_yaml/MANIFEST.in
@@ -1,0 +1,1 @@
+include VERSION
