+++ optional_plugins/loader_yaml/VERSION
@@ -1,0 +1,1 @@
+52.0
